day 5

+++ src/game/bases/Mathx.java
@@ -19,7 +19,7 @@ public class Mathx {
 //        return origin.multiply(1  - f).add(destination.multiply(f));
 //    }
 
-    public static double clamp(float value, float min, float max) {
+    public static float clamp(float value, float min, float max) {
         if (value < min) return min;
         if (value > max) return max;
         return value;
